refactor(cli-repl): use a object-oriented structure, hook loggers at initialization
I think this refactor provides much better readability and provides an easier interface for us to test and debug the functionality of this. This also moves the hookLoggers call to the point of initialization as
