+++ packages/logging/src/helpers.ts
